+++ templates/Uwp/Pages/MapVB/Views/MapPagePage.xaml.vb
@@ -5,5 +5,13 @@
         Public Sub New()
             InitializeComponent()
         End Sub
+
+        Protected Overrides Async Sub OnNavigatedTo(e As NavigationEventArgs)
+            Await ViewModel.InitializeAsync(mapControl)
+        End Sub
+
+        Protected Overrides Sub OnNavigatedFrom(e As NavigationEventArgs)
+            ViewModel.Cleanup()
+        End Sub
     End Class
 End Namespace
