--- templates/Uwp/_composition/CodeBehind/Page.Map.AddInitialization_Blank_SplitView/Views/wts.ItemNamePage_postaction.xaml.cs
@@ -1,24 +1,0 @@
-﻿//{[{
-using Windows.UI.Xaml.Navigation;
-//}]}
-namespace Param_ItemNamespace.Views
-{
-    public sealed partial class wts.ItemNamePage : Page, INotifyPropertyChanged
-    {
-        public wts.ItemNamePage()
-        {
-        }
-
-        //{[{
-        protected override async void OnNavigatedTo(NavigationEventArgs e)
-        {
-            await InitializeAsync();
-        }
-
-        protected override void OnNavigatedFrom(NavigationEventArgs e)
-        {
-            Cleanup();
-        }
-        //}]}
-    }
-}
