--- templates/Uwp/_composition/CodeBehind/Page.Map.AddInitialization_Blank_SplitViewVB/.template.config/template.json
@@ -1,29 +1,0 @@
-﻿{
-  "author": "Microsoft Community",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "CodeBehind.Page.Map.AddInitialization.BlankSplitView.VB",
-  "tags": {
-    "language": "VisualBasic",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform" : "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "identity == wts.Page.Map.CodeBehind.VB & $projectType == Blank|SplitView"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [
-  ],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
