--- templates/Uwp/_composition/_shared/Page.Map.AddInitialization_Blank_SplitView/Views/wts.ItemNamePage_postaction.xaml.cs
@@ -1,22 +1,0 @@
-﻿//{[{
-using Windows.UI.Xaml.Navigation;
-//}]}
-namespace Param_ItemNamespace.Views
-{
-    public sealed partial class wts.ItemNamePage : Page
-    {
-        //^^
-
-        //{[{
-        protected override async void OnNavigatedTo(NavigationEventArgs e)
-        {
-            await ViewModel.InitializeAsync(mapControl);
-        }
-
-        protected override void OnNavigatedFrom(NavigationEventArgs e)
-        {
-            ViewModel.Cleanup();
-        }
-        //}]}
-    }
-}
