--- templates/Uwp/_composition/_shared/Page.Map.AddInitialization_Blank_SplitViewVB/Views/wts.ItemNamePage_postaction.xaml.vb
@@ -1,16 +1,0 @@
-﻿Namespace Views
-    Public NotInheritable Partial Class wts.ItemNamePage
-        Inherits Page
-        '^^
-
-        '{[{
-        Protected Overrides Async Sub OnNavigatedTo(e As NavigationEventArgs)
-            Await ViewModel.InitializeAsync(mapControl)
-        End Sub
-
-        Protected Overrides Sub OnNavigatedFrom(e As NavigationEventArgs)
-            ViewModel.Cleanup()
-        End Sub
-        '}]}
-    End Class
-End Namespace
