--- templates/Uwp/_composition/_shared/Page.Map.AddInitialization_TabbedPivot/Views/wts.ItemNamePage_postaction.xaml.cs
@@ -1,29 +1,0 @@
-﻿//{[{
-using Windows.UI.Xaml;
-//}]}
-namespace Param_ItemNamespace.Views
-{
-    public sealed partial class wts.ItemNamePage : Page
-    {
-        public wts.ItemNamePage()
-        {
-            InitializeComponent();
-            //{[{
-            Loaded += wts.ItemNamePage_Loaded;
-            Unloaded += wts.ItemNamePage_Unloaded;
-            //}]}
-        }
-
-        //{[{
-        private async void wts.ItemNamePage_Loaded(object sender, RoutedEventArgs e)
-        {
-            await ViewModel.InitializeAsync(mapControl);
-        }
-
-        private void wts.ItemNamePage_Unloaded(object sender, RoutedEventArgs e)
-        {
-            ViewModel.Cleanup();
-        }
-        //}]}
-    }
-}
