--- templates/Uwp/_composition/_shared/Page.Map.AddInitialization_TabbedPivotVB/.template.config/template.json
@@ -1,29 +1,0 @@
-﻿{
-  "author": "Microsoft Community",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "Page.Map.AddInitialization_TabbedPivot.VB",
-  "tags": {
-    "language": "VisualBasic",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform" : "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == MVVMBasic|MVVMLight & $projectType == TabbedPivot & groupidentity == wts.Page.Map.VB"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [
-  ],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
